+++ Graph.h
@@ -77,7 +77,7 @@ public:
     //AGMG
     void auxPrimRandomizado(int initial_node, float alpha, int* parent, int* groups);
     void auxPrimGuloso(int initial_node, int* parent, int* groups);
-    float auxPrimReativo(float alpha, int max_iterations, int* doneiterations, float* weightFound);
+    float auxPrimReativo(float alpha, int max_iterations, int* doneiterations, double* weightFound);
     void primReativoAGMG(float* alpha, int alpha_size, int iterations, int block);
     float primRandomizadoAGMG(float alpha, int iterations);
     Graph* primGulosoAGMG();
